+++ eduVPN/ViewModels/Pages/ConnectionPage.cs
@@ -289,7 +289,7 @@ namespace eduVPN.ViewModels.Pages
                 try
                 {
                     using (var session =
-                        config.Protocol == VPNProtocol.WireGuard ? (Session)new WireGuardSession(Wizard, server, config.VPNConfig, expiration, config.ShouldFailover) :
+                        config.Protocol == VPNProtocol.WireGuard || config.Protocol == VPNProtocol.WireGuardProxy ? (Session)new WireGuardSession(Wizard, server, config.VPNConfig, expiration, config.ShouldFailover) :
                         config.Protocol == VPNProtocol.OpenVPN ? new OpenVPNSession(Wizard, server, config.VPNConfig, expiration, config.ShouldFailover) :
                             throw new ArgumentOutOfRangeException(nameof(config.Protocol), config.Protocol, null))
                     {