+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased](https://github.com/Amebis/eduVPN/compare/3.255.9...HEAD)
 
+- Support for WireGuard-over-TCP connections using Proxyguard
 - eduvpn-common updated to 987f2e3041e0b5d6e06cdc29657be99de9c1cd60
 - Preconfigurable start on sign-on